requirements.txt
openai>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
